backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
databases
python-jose[cryptography]
bcrypt ==4.0.1
passlib
PyJWT
python-decouple
python-multipart
requests
google-genai
email-validator
dotenv
polyline
googletrans
groq